// src/app/interfaces/contact.ts
export interface Contact {

    name?: string;
    lastName?: string;
    fullname?: string;
    phoneNumber: string;
    age?: number;
    email: string;
    company?: string;
    homepage?: string;
    family?: {
        sister?: string;
        brother?: string;
        dad?: string;
        mom?: string;
    }
    note?: string;
    __v?: number;
    _id?: string;
    id?: number;



}